requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-filter==25.1
pytest==8.4.2
pytest-django==4.11.1
sqlparse==0.5.3
asgiref==3.9.1
tzdata==2025.2
flake8==7.0.0
django-debug-toolbar==4.4.0
